cpuquiet: rqbalance: Add PM notifiers and online CPUs at resume.
This is for performance issues after system being suspended for
some time: on OSes like Android and/or userspace that will do its
synchronization, we experience sluggishness for 3 to 8 seconds,
depending on the initial load of the blocking queues.

Online CPUs and wait for the first worker complete iteration to
take over for offlining unneeded ones.

Signed-off-by: Luis <lefelite@gmail.com>